Added controller validations

+++ src/main/java/com/blueoptima/notes/controllers/NoteController.java
@@ -3,8 +3,10 @@ package com.blueoptima.notes.controllers;
 import com.blueoptima.notes.dao.NoteDao;
 import com.blueoptima.notes.models.Note;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -26,13 +28,24 @@ public class NoteController {
     public void createNote(@Valid @RequestBody Note note) { noteDao.createNote(note); }
 
     @RequestMapping(path = "api/notes/{noteId}", method = RequestMethod.PUT)
-    public void updateNote(@Valid @PathVariable int noteId, @RequestBody Note note) { noteDao.updateNote(noteId, note); }
+    public void updateNote(@Valid @PathVariable int noteId, @RequestBody Note note) {
+        String title = note.getTitle();
+        String content = note.getContent();
+
+        if (noteDao.getNote(noteId) == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Can't modify note: Note with ID: " + noteId + " not found");
+        if (title == null || title.isEmpty() || title.trim().isEmpty()) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't modify note: The title of the note must not be empty");
+        if (content == null || content.isEmpty() || content.trim().isEmpty()) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't modify note: The content of the note must not be empty");
+
+        noteDao.updateNote(noteId, note);
+    }
 
     @RequestMapping(path = "api/bulkUpdate/notes", method = RequestMethod.PUT)
     public void updateNotes() { noteDao.updateAllNotes(); }
 
     @RequestMapping(path = "api/notes/{noteId}", method = RequestMethod.DELETE)
     public void deleteNote(@PathVariable int noteId) {
+        if (noteDao.getNote(noteId) == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Can't delete note: Note with ID: " + noteId + " not found");
+
         noteDao.deleteNote(noteId);
     }
